feat: remove deprecated rules
